requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
openai
langchain
pandas
numpy
python-dotenv
telegram
scipy
CairoSVG
jupyter
faiss-cpu